refactor(fretboard): merge FretboardInteractive tests into Fretboard
- Add userEvent import to Fretboard.test.tsx
- Merge 3 interaction tests (click, drag, zoom resize) into new block
- Delete src/__tests__/FretboardInteractive.test.tsx
- Zero FretboardInteractive references remain in src/
